Do not create universal wheel
The universal wheel prevents the binary extension to be installed on
where not platform specific one is present. With this commit the a
wheel is created for PyPy 3.6, 3.7 and any implementation of Python 2,
as the extension is either not supported (Python 2) or decreases speed
(PyPy).
Pip will only now select this wheel for those Python versions, and tries
to build the extension where binary wheels are not available.